dags/alert_airflow_project/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.6
pandas>=1.5.0
apache-airflow>=2.5.0
